third_party/WebKit/LayoutTests/fast/scrolling/custom-scrollbar-style-applied.html

<!DOCTYPE HTML>
<script src="../../resources/js-test.js"></script>
<style>
::-webkit-scrollbar {
    width: 250px;
    height: 250px;
}
::-webkit-scrollbar-track {
    -webkit-box-shadow: inset 0 0 6px rgba(0,0,0,0.3); 
    -webkit-border-radius: 10px;
    border-radius: 10px;
}
::-webkit-scrollbar-thumb {
    -webkit-border-radius: 1px;
    background: rgba(255,0,0,0.8);
    -webkit-box-shadow: inset 0 0 1px rgba(255,250,0,0.5); 
}
#fill {
    background-color: #ccd;
    left: 0;
    top: 0;
    right: 0;
    bottom: 0;
    position: absolute;
}
#space {
    width: 1000px;
    height: 1000px;
}
</style>
<div id="fill"></div>
<div id="space"></div>
<script>
  var fill = document.querySelector("#fill");
  description("Tests that Styles are applied to custom scrollbar.");
  shouldBe("fill.clientWidth", "550");
  shouldBe("fill.clientHeight", "350");
</script>
