third_party/blink/web_tests/platform/generic/fast/scrolling/custom-scrollbar-style-applied-expected.txt

Tests that Styles are applied to custom scrollbar.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS fill.clientWidth is 550
PASS fill.clientHeight is 350
PASS successfullyParsed is true

TEST COMPLETE
